pyproject.toml
[tool.poetry]
name = "gfmrag"
version = "0.2.1"
description = "Graph Foundation Model for Retrieval Augmented Generation"
authors = [
    "Linhao Luo <linhao.luo@monash.edu>",
    "Zicheng Zhao <zicheng.zhao@njust.edu.cn>"
]
readme = "README.md"
packages = [{include = "gfmrag"}]

[tool.poetry.dependencies]
python = ">=3.12,<3.13"
torch = ">=2.4.1"
torch-geometric = "^2.4.0"
ninja = "^1.11.1.1"
easydict = "^1.13"
pyyaml = "^6.0.2"
tqdm = "^4.66.5"
sentence-transformers = ">=3.2.0"
hydra-core = "^1.3.2"
python-dotenv = "^1.0.1"
wandb = "^0.18.5"
transformers = "^4.46.1"
openai = "^1.52.2"
tiktoken = "^0.8.0"
langchain = "^0.3.9"
langchain-openai = "^0.2.11"
langchain-together = "^0.2.0"
langchain-community = "^0.3.9"
langchain-nvidia-ai-endpoints = "^0.3.9"
faiss-gpu-cu12 = "^1.9.0.post1"
ragatouille = "<0.0.10"


[tool.poetry.group.dev.dependencies]
pre-commit = "^4.0.1"
types-pyyaml = "^6.0.12.20240917"
mypy = "^1.12.1"
pytest = "^8.3.3"


[tool.poetry.group.doc.dependencies]
mkdocs = "^1.6.1"
mkdocstrings = "^0.27.0"
mkdocstrings-python = "^1.13.0"
mkdocs-autorefs = "^1.3.0"
mkdocs-material = "^9.5.50"
mkdocs-same-dir = "^0.1.3"
mike = "^2.1.3"

[tool.ruff]
exclude = ['gfmrag/ultra']

[tool.ruff.lint]
extend-select = ["I", "N", "UP", "B", "PD002"]
extend-ignore = ["E501", "E203", "E266"]
unfixable = ["B"]

[tool.ruff.lint.per-file-ignores]
"__init__.py" =  ["F401"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
